+++ MovieApp/src/main/java/com/KamilIsmail/MovieApp/repository/UserRepository.java
@@ -5,9 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
 
     List<UserEntity> findByUsername(String username);
-
-    UserEntity findByUserId(long id);
+    UserEntity findByUserId(int id);
 }
